requirements.txt
streamlit
schedule
python-dotenv
requests
pyyaml
flask
yfinance
matplotlib
pandas
